Move logging objects into _internal/logging.py
